pyproject.toml
[tool.poetry]
name = "gerdarzh-backend"
version = "0.1.0"
description = ""
authors = ["Benjamin D. <57868057+riwalld@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-cors-headers = "^4.6.0"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.6"
gunicorn = "^23.0.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
